#54 folders should be items as well

+++ src/app/components/entity-manager/entity-manager.component.ts
@@ -2,7 +2,7 @@ import { Component, Injector, OnInit, ViewChild } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute } from '@angular/router';
 import { EditorListComponent } from 'src/app/components/editor-list/editor-list.component';
-import { Entity } from 'src/app/entities/models/entity';
+import { Item } from 'src/app/entities/models/item';
 import { EntityService } from 'src/app/entities/services/entity.service';
 import { EditorMode } from '../editor/editor.component';
 import { PromptService } from '../prompts/prompt.service';
@@ -26,7 +26,7 @@ export class EntityManagerComponent implements OnInit {
     injector: Injector,
     private dialog: MatDialog,
     private promptService: PromptService,
-    ) {
+  ) {
     const entityServiceToken = route.snapshot.data["entityServiceToken"];
     this.editorMode = route.snapshot.data["editorMode"] ?? 'default';
     this.entityService = injector.get<EntityService>(entityServiceToken);
@@ -46,54 +46,68 @@ export class EntityManagerComponent implements OnInit {
   //#region public methods
   async createNewItem() {
     await this.editorListComponent.createTopLevelEntity();
-    this.refreshEntities();
+    this.refreshItems();
   }
 
   edit(path: string) {
     this.editorListComponent.openEditor(path);
   }
 
-  async rename(path: string) {
-    let entity: Entity | null = null;
-    const openedEditor = this.editorListComponent.editorComponents.find(x => x.entity.name === path);
-    if (openedEditor) {
-      entity = openedEditor?.entity;
-    } else {
-      entity = this.entityService.get(path);
-    }
+  async rename(oldPath: string, newPath: string = '') {
+    const openedEditor = this.editorListComponent.editorComponents.find(x => x.entity.path === oldPath);
 
-    if (!entity) {
-      return;
-    }
-
-    const initialName = entity.name;
-    const nameSegments = entity.name.split('/');
-    const newName = await this.promptService.openInputPrompt(this.dialog, 'New Name', nameSegments[nameSegments.length - 1]);
-    if (newName) {
-      nameSegments[nameSegments.length - 1] = newName;
-      entity.name = nameSegments.join('/');
-      if (openedEditor) {
-        openedEditor.setName(entity.name);
-        openedEditor.save();
+    if (!newPath) {
+      const itemPathSegments = oldPath.split('/');
+      const oldItemName = itemPathSegments[itemPathSegments.length - 1];
+      const newItemName = await this.promptService.openInputPrompt(this.dialog, 'New Name', oldItemName);
+      if (newItemName) {
+        itemPathSegments[itemPathSegments.length - 1] = newItemName;
+        newPath = itemPathSegments.join('/');
       } else {
-        this.entityService.create(entity.name, entity.rawContent);
+        return;
       }
-      this.entityService.delete(initialName);
-      this.refreshEntities();
+    }
+
+    if (newPath) {
+      if (openedEditor) {
+        openedEditor.setName(newPath);
+      }
+      this.entityService.move(oldPath, newPath);
+
+      const children = this.entityService.getDescendantsRecursive(oldPath);
+      for (const child of children) {
+        const newChildPath = child.path.replace(oldPath, newPath);
+        this.rename(child.path, newChildPath)
+      }
+
+      this.refreshItems();
     }
   }
 
-  delete(path: string) {
-    const openedEditor = this.editorListComponent.editorComponents.find(x => x.entity.name === path);
+  delete(path: string, confirmDeletion: boolean = true, deleteDescendants: boolean = true) {
+    if (confirmDeletion && !confirm(`Are you sure you want to delete ${path} and all its descendants?`)) {
+      return;
+    }
+
+    const openedEditor = this.editorListComponent.editorComponents.find(x => x.entity.path === path);
     if (openedEditor) {
       openedEditor.closeEditor();
     }
+
     this.entityService.delete(path);
+    this.refreshItems();
+
+    if (deleteDescendants) {
+      const descendants = this.entityService.getDescendantsRecursive(path);
+      for (const descendant of descendants) {
+        this.delete(descendant.path, false, false);
+      }
+    }
   }
   //#endregion
 
   //#region private methods
-  private refreshEntities() {
+  private refreshItems() {
     this.getAndSetTreeEntities();
     this.treeComponent.refreshItems(this.treeEntities);
   }
